requirements.txt
# Python libraries required for slowbeast
numpy
z3-solver
# Required to build llvmlite locally.
setuptools
# To setup unit tests
pytest
PyYAML